+++ lib/piper/command/ast/bind/option.ex
@@ -1,7 +1,7 @@
-defimpl Piper.Command.Bindable, for: Piper.Command.Ast2.Option do
+defimpl Piper.Command.Bindable, for: Piper.Command.Ast.Option do
 
   alias Piper.Command.Bindable
-  alias Piper.Command.Ast2
+  alias Piper.Command.Ast
 
   def resolve(option, scope) do
     if option.value != nil do
