+++ lib/piper/command/ast/chars/name_chars.ex
@@ -1,7 +1,10 @@
 defimpl String.Chars, for: Piper.Command.Ast.Name do
   alias Piper.Command.Ast.Name
 
-  def to_string(%Name{name: name}) do
-    "#{name}"
+  def to_string(%Name{bundle: nil, entity: entity}) do
+    "#{entity}"
+  end
+  def to_string(%Name{bundle: bundle, entity: entity}) do
+    "#{bundle}:#{entity}"
   end
 end
