+++ lib/piper/command/ast/chars/option_chars.ex
@@ -1,26 +1,18 @@
 defimpl String.Chars, for: Piper.Command.Ast.Option do
 
-  use Piper.Command.Ast.Chars.Util
   alias Piper.Command.Ast.Option
 
-  def to_string(%Option{flag: flag, value: nil}) do
-    flag = "#{flag}"
-    case String.length(flag) == 1 do
-      true ->
-        "-#{flag}"
-      false ->
-        "--#{flag}"
-    end
+  def to_string(%Option{name: name, value: nil, opt_type: :short}) do
+    "-#{name}"
   end
-
-  def to_string(%Option{flag: flag, value: value}) do
-    flag = "#{flag}"
-    case String.length(flag) == 1 do
-      true ->
-        "-#{flag}=#{escape(value)}"
-      false ->
-        "--#{flag}=#{escape(value)}"
-    end
+  def to_string(%Option{name: name, value: value, opt_type: :short}) do
+    "-#{name} #{value}"
+  end
+  def to_string(%Option{name: name, value: nil, opt_type: :long}) do
+    "--#{name}"
+  end
+  def to_string(%Option{name: name, value: value, opt_type: :long}) do
+    "--#{name}=#{value}"
   end
 
 end
