+++ lib/piper/command/ast/chars/variable_chars.ex
@@ -2,10 +2,11 @@ defimpl String.Chars, for: Piper.Command.Ast.Variable do
 
   alias Piper.Command.Ast.Variable
 
-  def to_string(%Variable{name: name, index: nil}) do
+  def to_string(%Variable{name: name, value: nil}) do
     "$#{name}"
   end
-  def to_string(%Variable{name: name, index: index}) do
-    "$#{name}[#{index}]"
+  def to_string(%Variable{value: value}) do
+    "#{value}"
   end
+
 end
