+++ lib/piper/command/ast/pipeline.ex
@@ -1,24 +1,11 @@
 defmodule Piper.Command.Ast.Pipeline do
 
   alias Piper.Command.Ast
-  alias Piper.Util.Token
 
-  defstruct [:line, :col, :finalized, :text, :type, :invocations, :output_to_input, :abort_on_error]
+  defstruct [:head]
 
-  def new(%Token{type: type}=token) when type in [:iff, :pipe] do
-    %__MODULE__{line: token.line, col: token.col,
-                text: token.text, type: token.type,
-                invocations: [],
-                output_to_input: (token.type == :pipe),
-                abort_on_error: true}
-  end
-
-  def add_invocation(%__MODULE__{invocations: invocations}=pipeline, %Ast.Invocation{}=invocation) do
-    %{pipeline | invocations: invocations ++ [invocation]}
-  end
-
-  def add_pipeline(%__MODULE__{invocations: invocations}=pipeline, %__MODULE__{}=next) do
-    %{pipeline | invocations: invocations ++ [next]}
+  def new(chain) do
+    %__MODULE__{head: chain}
   end
 
 end
