+++ lib/piper/command/ast/variable.ex
@@ -1,16 +1,9 @@
 defmodule Piper.Command.Ast.Variable do
 
-  use Piper.Util.TokenWrapper
+  defstruct [:line, :col, :name, :value]
 
-  defwrapper [value: :name, token_type: [:variable, :optvar],
-              others: [:index, :value, :binding_hook]]
-
-  def set_index(%__MODULE__{}=var, index) do
-    %{var | index: index}
-  end
-
-  def set_binding_hook(%__MODULE__{}=var, hook) when is_function(hook) do
-    %{var | binding_hook: hook}
+  def new({:variable, {line, col}, name}) do
+    %__MODULE__{line: line, col: col, name: name}
   end
 
 end
