--- lib/piper/command/ast2/chars/variable_chars.ex
@@ -1,12 +1,0 @@
-defimpl String.Chars, for: Piper.Command.Ast2.Variable do
-
-  alias Piper.Command.Ast2.Variable
-
-  def to_string(%Variable{name: name, value: nil}) do
-    "$#{name}"
-  end
-  def to_string(%Variable{value: value}) do
-    "#{value}"
-  end
-
-end
